requirements.txt
flask
requests
python-dotenv
mysql-connector-python
cryptography
boto3
gunicorn
prometheus_client